+++ changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Last change: 25-FEB-2026 CEST
+Last change: 28-FEB-2026 CEST
 
 0.9.8
 -----
@@ -9,6 +9,7 @@ Last change: 25-FEB-2026 CEST
 - ``PyImarisWriter`` is included in the one-click-installer again (Windows only)
 - Localize Parameters dialog displays a message if the z calibration path in the config file could not be found
 - Fixed Picasso Server launching in one-click-installers
+- ``picasso.localize.fit`` default method changed to ``sigmaxy`` (anisotropic sigma fitting)
 .. - Fixed 2D MLE fitting and cleaned the docstrings for better readability (``picasso.gaussmle``)
 
 0.9.7